refactor: 메인 배너 서버 컴포넌트로 변경
