+++ app/(route)/main/_components/Banner/index.tsx
@@ -1,13 +1,28 @@
+import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
+
+import getActivities from '@/_libs/activities/activitiesApi';
+
 import Carousel from '@/_components/Carousel';
 
-/**
- * Banner 컴포넌트 입니다.
- */
+export default async function Banner() {
+  const queryClient = new QueryClient();
 
-export default function Banner() {
+  await queryClient.prefetchQuery({
+    queryKey: ['main-banner'],
+    queryFn: () =>
+      getActivities({
+        method: 'cursor',
+        cursorId: null,
+        size: 3,
+        sort: 'most_reviewed',
+      }),
+  });
+
   return (
-    <div className="relative h-[240px] w-full mobile:h-[540px]">
-      <Carousel />
-    </div>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <div className="relative h-[240px] w-full mobile:h-[540px]">
+        <Carousel />
+      </div>
+    </HydrationBoundary>
   );
 }
