+++ app/_components/Carousel/index.tsx
@@ -8,28 +8,32 @@ import Autoplay from 'embla-carousel-autoplay';
 import Fade from 'embla-carousel-fade';
 import useEmblaCarousel from 'embla-carousel-react';
 import Image from 'next/image';
+import { useQuery } from '@tanstack/react-query';
 
-import useGetActivities from '@/_hooks/activities/useGetActivities';
+import getActivities from '@/_libs/activities/activitiesApi';
 
 import useCarouselDotBtn, { DotButton } from './CarouselDotBtn';
 
 import Btn from 'public/assets/icons/carousel-btn.svg';
-import Spinner from 'public/assets/icons/spinner.svg';
 
 const calendarNum = new Date().getMonth() + 1;
 
 /**
- * Carousel 컴포넌트 입니다.
+ * Main Banner Carousel 컴포넌트 입니다.
  */
 
 export default function Carousel() {
   const [emblaRef, emblaApi] = useEmblaCarousel({ loop: true, duration: 30 }, [Fade(), Autoplay({ delay: 3000 })]);
 
-  const { data, isLoading, isError } = useGetActivities({
-    method: 'cursor',
-    cursorId: null,
-    size: 3,
-    sort: 'most_reviewed',
+  const { data, isLoading, isError } = useQuery({
+    queryKey: ['main-banner'],
+    queryFn: () =>
+      getActivities({
+        method: 'cursor',
+        cursorId: null,
+        size: 3,
+        sort: 'most_reviewed',
+      }),
   });
 
   const scrollPrev = useCallback(() => {
@@ -56,7 +60,7 @@ export default function Carousel() {
   const { selectedIndex, scrollSnaps, onDotButtonClick } = useCarouselDotBtn(emblaApi, onNavButtonClick);
 
   if (isLoading) {
-    return <Image src={Spinner} fill alt="로딩중" className="size-20 mobile:size-32" />;
+    return <div className="size-full animate-pulse bg-gray-150" />;
   }
 
   if (isError) {
